backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
werkzeug==3.0.1
PyJWT==2.8.0
gunicorn==22.0.0
pytest==7.4.3
flask-jwt-extended==4.5.3
bcrypt==4.0.1
eventlet==0.35.2
